Refactor observers and make sure that there are no context exceptions
