+++ src/Model/Observer/CatalogLastPageRedirect.php
@@ -9,13 +9,46 @@
 namespace Emico\Tweakwise\Model\Observer;
 
 use Magento\Framework\Event\Observer;
+use Emico\Tweakwise\Model\Catalog\Layer\NavigationContext;
+use Emico\Tweakwise\Model\Config;
+use Magento\Framework\App\Action\Context;
+use Magento\Framework\Event\ObserverInterface;
+use Magento\Framework\HTTP\PhpEnvironment\Response;
 
-class CatalogLastPageRedirect extends AbstractProductNavigationRequestObserver
+class CatalogLastPageRedirect implements ObserverInterface
 {
+    /**
+     * @var Config
+     */
+    protected $config;
+
+    /**
+     * @var NavigationContext
+     */
+    protected $context;
+
+    /**
+     * @var Context
+     */
+    private $actionContext;
+
+    /**
+     * CatalogSearchRedirect constructor.
+     * @param Config $config
+     * @param NavigationContext $context
+     * @param Context $actionContext
+     */
+    public function __construct(Config $config, NavigationContext $context, Context $actionContext)
+    {
+        $this->config = $config;
+        $this->context = $context;
+        $this->actionContext = $actionContext;
+    }
+
     /**
      * {@inheritdoc}
      */
-    protected function _execute(Observer $observer)
+    public function execute(Observer $observer)
     {
         $properties = $this->context->getResponse()->getProperties();
         if (!$properties->getNumberOfItems()) {
@@ -28,12 +61,30 @@ class CatalogLastPageRedirect extends AbstractProductNavigationRequestObserver
             return;
         }
 
-        $url = $this->urlBuilder->getUrl('*/*/*', [
+        $response = $this->getHttpResponse();
+        if (!$response || $response->isRedirect()) {
+            return;
+        }
+
+        $url = $this->actionContext->getUrl()->getUrl('*/*/*', [
             '_current' => true,
             '_use_rewrite' => true,
             '_query' => ['p' => $lastPage]
         ]);
 
-        $this->getHttpResponse()->setRedirect($url);
+        $response->setRedirect($url);
+    }
+
+    /**
+     * @return Response|null
+     */
+    protected function getHttpResponse()
+    {
+        $response = $this->actionContext->getResponse();
+        if (!$response instanceof Response) {
+            return null;
+        }
+
+        return $response;
     }
 }
