+++ src/Model/Observer/CatalogSearchLastPageRedirect.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * @author : Edwin Jacobs, email: ejacobs@emico.nl.
+ * @copyright : Copyright Emico B.V. 2019.
+ */
+
+namespace Emico\Tweakwise\Model\Observer;
+
+use Magento\Framework\Event\Observer;
+
+class CatalogSearchLastPageRedirect extends CatalogLastPageRedirect
+{
+    /**
+     * @param Observer $observer
+     */
+    public function execute(Observer $observer)
+    {
+        if ($this->config->isSearchEnabled()) {
+            parent::execute($observer);
+        }
+    }
+}
