+++ src/etc/di.xml
@@ -76,9 +76,24 @@
     <type name="Emico\Tweakwise\Model\Observer\CatalogSearchRedirect">
         <arguments>
             <argument name="config" xsi:type="object">Emico\Tweakwise\Model\Config</argument>
-            <argument name="context" xsi:type="object">Emico\Tweakwise\Model\Catalog\Layer\NavigationContext\CurrentContext</argument>
+            <argument name="context" xsi:type="object">Emico\Tweakwise\Model\Catalog\Layer\NavigationContext\Search</argument>
             <argument name="actionContext" xsi:type="object">Magento\Framework\App\Action\Context</argument>
-            <argument name="navigationContext" xsi:type="object">Emico\Tweakwise\Model\Catalog\Layer\NavigationContext\Search</argument>
+        </arguments>
+    </type>
+
+    <type name="Emico\Tweakwise\Model\Observer\CatalogSearchLastPageRedirect">
+        <arguments>
+            <argument name="config" xsi:type="object">Emico\Tweakwise\Model\Config</argument>
+            <argument name="context" xsi:type="object">Emico\Tweakwise\Model\Catalog\Layer\NavigationContext\Search</argument>
+            <argument name="actionContext" xsi:type="object">Magento\Framework\App\Action\Context</argument>
+        </arguments>
+    </type>
+
+    <type name="Emico\Tweakwise\Model\Observer\CatalogNavigationLastPageRedirect">
+        <arguments>
+            <argument name="config" xsi:type="object">Emico\Tweakwise\Model\Config</argument>
+            <argument name="context" xsi:type="object">Emico\Tweakwise\Model\Catalog\Layer\NavigationContext\Category</argument>
+            <argument name="actionContext" xsi:type="object">Magento\Framework\App\Action\Context</argument>
         </arguments>
     </type>
